+++ blog/src/screens/EditScreen.js
@@ -1,6 +1,7 @@
-import React, { useState, useContext } from 'react';
-import { View, Text, StyleSheet, TextInput } from 'react-native';
+import React, { useContext } from 'react';
+import { StyleSheet } from 'react-native';
 import { Context } from '../contetx/BlogContext';
+import BlogPostForm from '../components/BlogPostForm';
 
 const EditScreen = ({ navigation }) => {
   const { state } = useContext(Context);
@@ -9,19 +10,7 @@ const EditScreen = ({ navigation }) => {
     blogPost => blogPost.id === navigation.getParam('id')
   );
 
-  const [title, setTitle] = useState(blogPost.title);
-  const [content, setContent] = useState(blogPost.content);
-  return (
-    <View>
-      <Text>Edit title:</Text>
-      <TextInput value={title} onChangeText={newTitle => setTitle(newTitle)} />
-      <Text>Edit content:</Text>
-      <TextInput
-        value={content}
-        onChangeText={newContent => setContent(newContent)}
-      />
-    </View>
-  );
+  return <BlogPostForm />;
 };
 const styles = StyleSheet.create({});
